main.go: moves up the entry point to the root. moves the packages for individual commands under the `internal` directory.
