+++ internal/cmd/register/register.go
